Combined PresetTimerRequestModel and CustomTimerRequestModel into TimerRequestModel Created MainFrame.java and MainFrame.form for the UI Fixed timer countdown Refactored package names and organized directory
