+++ src/main/java/Timer/timer_screen/CustomTimerTextField.java
@@ -1,4 +1,4 @@
-package Timer.TimerUI;
+package Timer.timer_screen;
 
 public class CustomTimerTextField {
 }
