+++ src/main/java/Timer/timer_screen/PresetTimerButtons.java
@@ -1,4 +1,4 @@
-package Timer.TimerUI;
+package Timer.timer_screen;
 
 import javax.swing.*;
