+++ src/main/java/Timer/timer_use_cases/ProgressBarInteractor.java
@@ -1,4 +1,4 @@
-package Timer.TimerUseCases;
+package Timer.timer_use_cases;
 
 public class ProgressBarInteractor {
     private void resetProgressBar() {
