+++ src/main/java/Timer/timer_use_cases/ProgressBarOutputBoundary.java
@@ -1,4 +1,4 @@
-package Timer.TimerOutputBoundaries;
+package Timer.timer_use_cases;
 
 public interface ProgressBarOutputBoundary {
     void displayProgressBar();
