+++ src/main/java/Timer/timer_use_cases/TimerInputBoundary.java
@@ -1,4 +1,4 @@
-package Timer.TimerInputBoundaries;
+package Timer.timer_use_cases;
 
 public interface TimerInputBoundary {
     public void startTimer();
