+++ src/main/java/Timer/timer_use_cases/TimerResponseModel.java
@@ -1,4 +1,4 @@
-package Timer.TimerUseCases;
+package Timer.timer_use_cases;
 
 public class TimerResponseModel {
     private String outputTime = "00:00:00";
